Refactor: Split large test files into smaller components
I've addressed issue #<ISSUE_ID_IF_AVAILABLE> by breaking down several large test files into smaller, more focused modules. The goal is to improve maintainability and potentially mitigate issues like Out-of-Memory errors or hangs during test execution.

The following major test files were split:
- src/tests/contexts/EarthquakeDataContext.test.jsx (into reducer, provider, and selector tests)
- functions/[[catchall]].test.js (into tests for USGS proxy, sitemaps, and prerendering routes)
- src/utils/utils.test.js (into individual test files for each utility function)
- src/components/EarthquakeDetailModalComponent.test.jsx (into SEO, navigation, data, and common tests)
- functions/api/calculate-clusters.test.js (into endpoint and internal logic tests)

NOTE: Due to a persistent issue within the test execution environment, I was unsuccessful in running `npm test` after these file system modifications. The command failed with an error related to computing affected file counts, preventing verification of the test suite's status or observation of OOM/hanging issues. The file splitting has been completed as requested.
