+++ src/tests/contexts/EarthquakeDataContext.test.jsx
@@ -1,278 +1,40 @@
-import React from 'react';
-import { EarthquakeDataProvider, useEarthquakeDataState } from '../../contexts/EarthquakeDataContext';
-import { earthquakeReducer, initialState, actionTypes, initialState as contextInitialState, EarthquakeDataContext } from '../../contexts/earthquakeDataContextUtils.js'; // Corrected path and consolidated initialState, added EarthquakeDataContext
+// This file will now primarily contain tests for helper functions, if any,
+// or could be a placeholder if all tests are moved.
+
+// import React from 'react'; // Keep if helper tests might need it or for consistency
+// import { EarthquakeDataProvider, useEarthquakeDataState } from '../../contexts/EarthquakeDataContext'; // Likely not needed
+// import { earthquakeReducer, initialState, actionTypes, initialState as contextInitialState, EarthquakeDataContext } from '../../contexts/earthquakeDataContextUtils.js'; // Likely not needed
 
 // --- React specific testing imports ---
-import { renderHook, act, waitFor } from '@testing-library/react';
-import { vi } from 'vitest';
-// contextActionTypes removed from import as it's unused
-// import { initialState as contextInitialState, useEarthquakeDataState } from '../../contexts/EarthquakeDataContext'; // This initialState might need to change too if it's the same one.
-import { fetchUsgsData } from '../../services/usgsApiService';
-import {
-    FEELABLE_QUAKE_THRESHOLD,
-    MAJOR_QUAKE_THRESHOLD,
-    USGS_API_URL_MONTH,
-    USGS_API_URL_DAY,
-    USGS_API_URL_WEEK,
-    LOADING_MESSAGE_INTERVAL_MS,
-    REFRESH_INTERVAL_MS, // Added import
-    // Constants needed for reducer tests if not available in main initialState/actionTypes
-    // For example, if MAJOR_QUAKE_THRESHOLD is used directly in reducer tests:
-    // MAJOR_QUAKE_THRESHOLD as APP_MAJOR_QUAKE_THRESHOLD
-} from '../../constants/appConstants';
+// import { renderHook, act, waitFor } from '@testing-library/react'; // Likely not needed
+// import { vi } from 'vitest'; // Likely not needed if mocks are specific to other files
+// import { fetchUsgsData } from '../../services/usgsApiService'; // Likely not needed
 
-// Mock the usgsApiService
-vi.mock('../../services/usgsApiService', () => ({
-  fetchUsgsData: vi.fn(),
-}));
+// Import constants only if they are directly used by helper functions tested in this file.
+// import {
+//     FEELABLE_QUAKE_THRESHOLD,
+//     MAJOR_QUAKE_THRESHOLD,
+//     USGS_API_URL_MONTH,
+//     USGS_API_URL_DAY,
+//     USGS_API_URL_WEEK,
+//     LOADING_MESSAGE_INTERVAL_MS,
+//     REFRESH_INTERVAL_MS,
+// } from '../../constants/appConstants';
+
+// Mock the usgsApiService - only if helper tests in this file would trigger it.
+// vi.mock('../../services/usgsApiService', () => ({
+//   fetchUsgsData: vi.fn(),
+// }));
 
 // --- Helper Function Definitions (copied from EarthquakeDataContext.jsx for isolated testing) ---
 // These are assumed to be correct and tested by their own suites below.
-const filterByTime = (data, hoursAgoStart, hoursAgoEnd = 0, now = Date.now()) => { if (!Array.isArray(data)) return []; const startTime = now - hoursAgoStart * 36e5; const endTime = now - hoursAgoEnd * 36e5; return data.filter(q => q.properties.time >= startTime && q.properties.time < endTime);};
-const filterMonthlyByTime = (data, daysAgoStart, daysAgoEnd = 0, now = Date.now()) => { if (!Array.isArray(data)) return []; const startTime = now - (daysAgoStart * 24 * 36e5); const endTime = now - (daysAgoEnd * 24 * 36e5); return data.filter(q => q.properties.time >= startTime && q.properties.time < endTime);};
-const consolidateMajorQuakesLogic = (currentLastMajor, currentPreviousMajor, newMajors) => { let potentialQuakes = [...newMajors]; if (currentLastMajor) potentialQuakes.push(currentLastMajor); if (currentPreviousMajor) potentialQuakes.push(currentPreviousMajor); const consolidated = potentialQuakes.sort((a, b) => b.properties.time - a.properties.time).filter((quake, index, self) => index === self.findIndex(q => q.id === quake.id)); const newLastMajor = consolidated.length > 0 ? consolidated[0] : null; const newPreviousMajor = consolidated.length > 1 ? consolidated[1] : null; const newTimeBetween = newLastMajor && newPreviousMajor ? newLastMajor.properties.time - newPreviousMajor.properties.time : null; return { lastMajorQuake: newLastMajor, previousMajorQuake: newPreviousMajor, timeBetweenPreviousMajorQuakes: newTimeBetween };};
-// const sampleArray = (array, sampleSize) => { if (!Array.isArray(array) || array.length === 0) return []; if (sampleSize <= 0) return []; if (sampleSize >= array.length) return [...array]; const shuffled = [...array]; for (let i = shuffled.length - 1; i > 0; i--) { const j = Math.floor(Math.random() * (i + 1)); [shuffled[i], shuffled[j]] = [shuffled[j], shuffled[i]]; } return shuffled.slice(0, sampleSize);}; // Unused function removed
-// function sampleArrayWithPriority (fullArray, sampleSize, priorityMagnitudeThreshold) { ... } // Removed unused helper
-const MAGNITUDE_RANGES = [ {name: '<1', min: -Infinity, max: 0.99}, {name : '1-1.9', min : 1, max : 1.99}, {name: '2-2.9', min: 2, max: 2.99}, {name : '3-3.9', min : 3, max : 3.99}, {name: '4-4.9', min: 4, max: 4.99}, {name : '5-5.9', min : 5, max : 5.99}, {name: '6-6.9', min: 6, max: 6.99}, {name : '7+', min : 7, max : Infinity}, ];
-// const formatDateForTimeline = (timestamp) => { ... }; // Removed, as getInitialDailyCounts which used it is also removed
-// const getInitialDailyCounts = (numDays, baseTime) => { ... }; // Removed unused helper
-// const getMagnitudeColor = (mag) => `color_for_mag_${mag}`; // Removed, as calculateMagnitudeDistribution which used it is also removed
-// const calculateMagnitudeDistribution = (earthquakes) => { ... }; // Removed unused helper
+// const filterByTime = (data, hoursAgoStart, hoursAgoEnd = 0, now = Date.now()) => { if (!Array.isArray(data)) return []; const startTime = now - hoursAgoStart * 36e5; const endTime = now - hoursAgoEnd * 36e5; return data.filter(q => q.properties.time >= startTime && q.properties.time < endTime);};
+// const filterMonthlyByTime = (data, daysAgoStart, daysAgoEnd = 0, now = Date.now()) => { if (!Array.isArray(data)) return []; const startTime = now - (daysAgoStart * 24 * 36e5); const endTime = now - (daysAgoEnd * 24 * 36e5); return data.filter(q => q.properties.time >= startTime && q.properties.time < endTime);};
+// const consolidateMajorQuakesLogic = (currentLastMajor, currentPreviousMajor, newMajors) => { let potentialQuakes = [...newMajors]; if (currentLastMajor) potentialQuakes.push(currentLastMajor); if (currentPreviousMajor) potentialQuakes.push(currentPreviousMajor); const consolidated = potentialQuakes.sort((a, b) => b.properties.time - a.properties.time).filter((quake, index, self) => index === self.findIndex(q => q.id === quake.id)); const newLastMajor = consolidated.length > 0 ? consolidated[0] : null; const newPreviousMajor = consolidated.length > 1 ? consolidated[1] : null; const newTimeBetween = newLastMajor && newPreviousMajor ? newLastMajor.properties.time - newPreviousMajor.properties.time : null; return { lastMajorQuake: newLastMajor, previousMajorQuake: newPreviousMajor, timeBetweenPreviousMajorQuakes: newTimeBetween };};
+// const MAGNITUDE_RANGES = [ {name: '<1', min: -Infinity, max: 0.99}, {name : '1-1.9', min : 1, max : 1.99}, {name: '2-2.9', min: 2, max: 2.99}, {name : '3-3.9', min : 3, max : 3.99}, {name: '4-4.9', min: 4, max: 4.99}, {name : '5-5.9', min : 5, max : 5.99}, {name: '6-6.9', min: 6, max: 6.99}, {name : '7+', min : 7, max : Infinity}, ];
 
 // --- Reducer Tests ---
-describe('EarthquakeDataContext Reducer', () => {
-  it('should return the initial state', () => expect(earthquakeReducer(undefined, {})).toEqual(initialState));
-
-  // Test SET_LOADING_FLAGS thoroughly
-  describe('SET_LOADING_FLAGS action', () => {
-    it('should update multiple loading flags', () => {
-      const payload = { isLoadingDaily: false, isLoadingWeekly: false, isLoadingMonthly: true };
-      const expectedState = { ...initialState, ...payload };
-      expect(earthquakeReducer(initialState, { type: actionTypes.SET_LOADING_FLAGS, payload })).toEqual(expectedState);
-    });
-    it('should update a single loading flag', () => {
-      const payload = { isLoadingMonthly: true };
-      const expectedState = { ...initialState, ...payload };
-      expect(earthquakeReducer(initialState, { type: actionTypes.SET_LOADING_FLAGS, payload })).toEqual(expectedState);
-    });
-     it('should not change other state properties', () => {
-      const payload = { isLoadingDaily: false };
-      const stateBefore = { ...initialState, earthquakesLastHour: [{ id: 'test' }] };
-      const stateAfter = earthquakeReducer(stateBefore, { type: actionTypes.SET_LOADING_FLAGS, payload });
-      expect(stateAfter.isLoadingDaily).toBe(false);
-      expect(stateAfter.earthquakesLastHour).toEqual([{ id: 'test' }]); // Ensure other parts are untouched
-    });
-  });
-
-  // Test SET_ERROR thoroughly
-  describe('SET_ERROR action', () => {
-    it('should update general error and clear monthly error', () => {
-      const payload = { error: 'General Error', monthlyError: null }; // Explicitly clearing monthlyError
-      const stateBefore = { ...initialState, monthlyError: 'Previous Monthly Error' };
-      const expectedState = { ...stateBefore, ...payload };
-      expect(earthquakeReducer(stateBefore, { type: actionTypes.SET_ERROR, payload })).toEqual(expectedState);
-    });
-    it('should update monthlyError and clear general error', () => {
-      const payload = { monthlyError: 'Monthly Error', error: null }; // Explicitly clearing general error
-      const stateBefore = { ...initialState, error: 'Previous General Error' };
-      const expectedState = { ...stateBefore, ...payload };
-      expect(earthquakeReducer(stateBefore, { type: actionTypes.SET_ERROR, payload })).toEqual(expectedState);
-    });
-    it('should update only one error if the other is undefined in payload', () => {
-      let payload = { error: 'New General Error' };
-      let stateBefore = { ...initialState, monthlyError: 'Existing Monthly Error' };
-      let expectedState = { ...stateBefore, error: 'New General Error' };
-      expect(earthquakeReducer(stateBefore, { type: actionTypes.SET_ERROR, payload })).toEqual(expectedState);
-
-      payload = { monthlyError: 'New Monthly Error' };
-      stateBefore = { ...initialState, error: 'Existing General Error' };
-      expectedState = { ...stateBefore, monthlyError: 'New Monthly Error' };
-      expect(earthquakeReducer(stateBefore, { type: actionTypes.SET_ERROR, payload })).toEqual(expectedState);
-    });
-  });
-
-  it('should handle SET_INITIAL_LOAD_COMPLETE', () => expect(earthquakeReducer(initialState, { type: actionTypes.SET_INITIAL_LOAD_COMPLETE })).toEqual({ ...initialState, isInitialAppLoad: false }));
-  it('should handle UPDATE_LOADING_MESSAGE_INDEX', () => { const s = {...initialState,currentLoadingMessages:['1','2'],loadingMessageIndex:0}; expect(earthquakeReducer(s,{type:actionTypes.UPDATE_LOADING_MESSAGE_INDEX}).loadingMessageIndex).toBe(1);});
-  it('should handle UPDATE_LOADING_MESSAGE_INDEX and cycle to 0', () => { const s = {...initialState,currentLoadingMessages:['1','2'],loadingMessageIndex:1}; expect(earthquakeReducer(s,{type:actionTypes.UPDATE_LOADING_MESSAGE_INDEX}).loadingMessageIndex).toBe(0);});
-  it('should handle SET_LOADING_MESSAGES', () => { const m = ['New']; expect(earthquakeReducer(initialState,{type:actionTypes.SET_LOADING_MESSAGES,payload:m})).toEqual({...initialState,currentLoadingMessages:m,loadingMessageIndex:0});});
-
-  describe('DAILY_DATA_PROCESSED action', () => {
-    const mockFetchTime = Date.now(); const mockMetadata = {generated:mockFetchTime-1000};
-    const createMockQuake = (id,offset,mag,alert='green',tsunami=0) => ({id,properties:{mag,time:mockFetchTime-offset*36e5,alert,tsunami}});
-    const mockFeatures = [createMockQuake('q1',0.5,2.5),createMockQuake('q2',1.5,3.0),createMockQuake('q3',10,4.5),createMockQuake('q4',30,5.5,'yellow'),createMockQuake('major1',2,6.0,'red',1)];
-    const action = {type:actionTypes.DAILY_DATA_PROCESSED,payload:{features:mockFeatures,metadata:mockMetadata,fetchTime:mockFetchTime}};
-    const updatedState = earthquakeReducer(initialState,action);
-    it('should set isLoadingDaily to false and update fetch times', () => {expect(updatedState.isLoadingDaily).toBe(false);expect(updatedState.dataFetchTime).toBe(mockFetchTime);expect(updatedState.lastUpdated).toBe(new Date(mockMetadata.generated).toLocaleString());});
-    it('should filter earthquakes', () => {expect(updatedState.earthquakesLastHour.length).toBe(1);expect(updatedState.earthquakesPriorHour.length).toBe(2);expect(updatedState.earthquakesLast24Hours.length).toBe(4);});
-    it('should update tsunami and alert status', () => {expect(updatedState.hasRecentTsunamiWarning).toBe(true);expect(updatedState.highestRecentAlert).toBe('red');expect(updatedState.activeAlertTriggeringQuakes.length).toBe(1);});
-    it('should consolidate major quakes', () => {expect(updatedState.lastMajorQuake.id).toBe('major1');expect(updatedState.previousMajorQuake.id).toBe('q3');});
-
-    it('should handle no alerts', () => {
-      const noAlertFeatures = [createMockQuake('q_noalert', 0.5, 2.5, 'green', 0)];
-      const actionNoAlert = {...action, payload: {...action.payload, features: noAlertFeatures}};
-      const stateNoAlert = earthquakeReducer(initialState, actionNoAlert);
-      expect(stateNoAlert.highestRecentAlert).toBeNull();
-      expect(stateNoAlert.activeAlertTriggeringQuakes.length).toBe(0);
-    });
-
-    it('should handle no major quakes', () => {
-      const noMajorFeatures = [createMockQuake('q_minor', 0.5, 2.5)];
-      const actionNoMajor = {...action, payload: {...action.payload, features: noMajorFeatures}};
-      const stateNoMajor = earthquakeReducer(initialState, actionNoMajor);
-      expect(stateNoMajor.lastMajorQuake).toBeNull();
-      expect(stateNoMajor.previousMajorQuake).toBeNull();
-    });
-  });
-
-  describe('WEEKLY_DATA_PROCESSED action', () => {
-    const mockFetchTime = Date.now();
-    const createMockQuake = (id, timeOffsetHours, mag) => ({ id, properties: { mag, time: mockFetchTime - timeOffsetHours * 36e5 } });
-    // Modified mockFeaturesWeekly with duplicate IDs
-    const mockFeaturesWeekly = [
-      createMockQuake('w_quake1', 10, 2.5),
-      createMockQuake('w_quake2', 50, 3.0),
-      createMockQuake('w_quake3', 80, 4.5), // This one is > 72h, will be filtered out by filterByTime
-      createMockQuake('w_quakeMajor', 60, 5.8), // Original major quake
-      createMockQuake('w_quakeMajor', 61, 5.7), // Duplicate ID, slightly different time & mag, within 72h
-      createMockQuake('w_anotherMajor', 30, 5.9),
-      createMockQuake('w_anotherMajor', 31, 5.85), // Duplicate ID for another major quake, within 72h
-      createMockQuake('w_highMagDup', 20, 6.0), // High magnitude quake
-      createMockQuake('w_highMagDup', 22, 5.9), // Duplicate ID for high magnitude quake
-      ...Array.from({ length: 5 }, (_, i) => createMockQuake(`w_extra${i}`, 24 * (i + 1) , 3.0 + i * 0.1)) // Keep some other quakes (some > 72h)
-    ];
-    const action = { type: actionTypes.WEEKLY_DATA_PROCESSED, payload: { features: mockFeaturesWeekly, fetchTime: mockFetchTime } };
-    const updatedState = earthquakeReducer(initialState, action);
-
-    it('should set isLoadingWeekly to false', () => expect(updatedState.isLoadingWeekly).toBe(false));
-
-    it('should filter earthquakes for last 72 hours (deduplicated) and last 7 days', () => {
-      // Test earthquakesLast72Hours (which should be deduplicated by the reducer)
-      const expectedIn72HoursRaw = mockFeaturesWeekly.filter(q => (mockFetchTime - q.properties.time) <= 72 * 36e5);
-      const uniqueIds72h = new Set();
-      const expectedDeduplicatedIn72Hours = expectedIn72HoursRaw.filter(quake => {
-          if (!uniqueIds72h.has(quake.id)) {
-              uniqueIds72h.add(quake.id);
-              return true;
-          }
-          return false;
-      });
-      expect(updatedState.earthquakesLast72Hours.map(q=>q.id).sort()).toEqual(expectedDeduplicatedIn72Hours.map(q=>q.id).sort());
-      expect(updatedState.earthquakesLast72Hours.length).toEqual(expectedDeduplicatedIn72Hours.length);
-      // Test earthquakesLast7Days (no deduplication applied here in the reducer)
-      // This assertion was previously checking against mockFeaturesWeekly.length which includes quakes beyond 7 days.
-      // It should filter to the actual 7-day window.
-      const expectedLast7Days = mockFeaturesWeekly.filter(q => (mockFetchTime - q.properties.time) <= 7 * 24 * 36e5);
-      expect(updatedState.earthquakesLast7Days.map(q=>q.id).sort()).toEqual(expectedLast7Days.map(q=>q.id).sort());
-    });
-    it('should populate globeEarthquakes (sorted subset of last 72 hours, deduplicated)', () => {
-        const last72HoursData = filterByTime(mockFeaturesWeekly, 72, 0, mockFetchTime);
-
-        // Deduplication step for the test's expectation
-        const uniqueEarthquakeIds = new Set();
-        const deduplicatedLast72HoursData = last72HoursData.filter(quake => {
-            if (!uniqueEarthquakeIds.has(quake.id)) {
-                uniqueEarthquakeIds.add(quake.id);
-                return true;
-            }
-            return false;
-        });
-
-        const expectedGlobeQuakes = [...deduplicatedLast72HoursData].sort((a,b) => (b.properties.mag || 0) - (a.properties.mag || 0)).slice(0, 900);
-        expect(updatedState.globeEarthquakes.map(q=>q.id).sort()).toEqual(expectedGlobeQuakes.map(q=>q.id).sort());
-        // Also check that the count is as expected after deduplication
-        expect(updatedState.globeEarthquakes.length).toEqual(expectedGlobeQuakes.length);
-    });
-    it('should filter prev24HourData (48h to 24h ago)', () => {
-        const expectedPrev24 = filterByTime(mockFeaturesWeekly, 48, 24, mockFetchTime);
-        expect(updatedState.prev24HourData.map(q=>q.id).sort()).toEqual(expectedPrev24.map(q=>q.id).sort());
-    });
-    it('should consolidate major quakes from weekly data', () => {
-        const weeklyMajors = mockFeaturesWeekly.filter(q => q.properties.mag !== null && q.properties.mag >= MAJOR_QUAKE_THRESHOLD);
-        const mjUpdates = consolidateMajorQuakesLogic(initialState.lastMajorQuake, initialState.previousMajorQuake, weeklyMajors);
-        expect(updatedState.lastMajorQuake?.id).toBe(mjUpdates.lastMajorQuake?.id);
-    });
-    it('should calculate dailyCounts7Days, sampledEarthquakesLast7Days, and magnitudeDistribution7Days', () => {
-        expect(updatedState.dailyCounts7Days.length).toBe(7);
-        expect(updatedState.sampledEarthquakesLast7Days).toBeInstanceOf(Array);
-        expect(updatedState.magnitudeDistribution7Days.length).toBe(MAGNITUDE_RANGES.length);
-    });
-
-    it('should handle empty features for weekly processing', () => {
-      const emptyAction = { type: actionTypes.WEEKLY_DATA_PROCESSED, payload: { features: [], fetchTime: mockFetchTime } };
-      const emptyState = earthquakeReducer(initialState, emptyAction);
-      expect(emptyState.isLoadingWeekly).toBe(false);
-      expect(emptyState.earthquakesLast72Hours.length).toBe(0);
-      expect(emptyState.earthquakesLast7Days.length).toBe(0);
-      expect(emptyState.globeEarthquakes.length).toBe(0);
-      expect(emptyState.dailyCounts7Days.every(d => d.count === 0)).toBe(true);
-      expect(emptyState.sampledEarthquakesLast7Days.length).toBe(0);
-      expect(emptyState.magnitudeDistribution7Days.every(d => d.count === 0)).toBe(true);
-      expect(emptyState.lastMajorQuake).toBeNull();
-    });
-  });
-
-  describe('MONTHLY_DATA_PROCESSED action', () => {
-    const mockFetchTime = Date.now();
-    const createMockQuake = (id, timeOffsetDays, mag) => ({ id, properties: { mag, time: mockFetchTime - timeOffsetDays * 24 * 36e5 } });
-    const mockFeaturesMonthly = [
-      createMockQuake('m_quake1', 5, 2.5), createMockQuake('m_quakeMajor', 15, 6.2),
-      ...Array.from({ length: 10 }, (_, i) => createMockQuake(`m_extra${i}`, i + 1 , 2.0 + i * 0.2))
-    ];
-    const action = { type: actionTypes.MONTHLY_DATA_PROCESSED, payload: { features: mockFeaturesMonthly, fetchTime: mockFetchTime } };
-    const updatedState = earthquakeReducer(initialState, action);
-
-    it('should set isLoadingMonthly to false, hasAttemptedMonthlyLoad to true', () => {
-        expect(updatedState.isLoadingMonthly).toBe(false);
-        expect(updatedState.hasAttemptedMonthlyLoad).toBe(true);
-    });
-    it('should populate allEarthquakes, earthquakesLast14Days, and earthquakesLast30Days', () => {
-        expect(updatedState.allEarthquakes.length).toBe(mockFeaturesMonthly.length);
-        const expected14day = filterMonthlyByTime(mockFeaturesMonthly, 14, 0, mockFetchTime);
-        expect(updatedState.earthquakesLast14Days.map(q=>q.id).sort()).toEqual(expected14day.map(q=>q.id).sort());
-    });
-    it('should filter prev7DayData and prev14DayData', () => {
-        const expectedP7 = filterMonthlyByTime(mockFeaturesMonthly, 14, 7, mockFetchTime);
-        const expectedP14 = filterMonthlyByTime(mockFeaturesMonthly, 28, 14, mockFetchTime);
-        expect(updatedState.prev7DayData.map(q=>q.id).sort()).toEqual(expectedP7.map(q=>q.id).sort());
-        expect(updatedState.prev14DayData.map(q=>q.id).sort()).toEqual(expectedP14.map(q=>q.id).sort());
-    });
-    it('should consolidate major quakes from monthly data', () => {
-        const monthlyMajors = mockFeaturesMonthly.filter(q => q.properties.mag !== null && q.properties.mag >= MAJOR_QUAKE_THRESHOLD);
-        const mjUpdates = consolidateMajorQuakesLogic(initialState.lastMajorQuake, initialState.previousMajorQuake, monthlyMajors);
-        expect(updatedState.lastMajorQuake?.id).toBe(mjUpdates.lastMajorQuake?.id);
-    });
-    it('should calculate dailyCounts, sampledEarthquakes, and magnitudeDistribution for 14/30 days', () => {
-        expect(updatedState.dailyCounts14Days.length).toBe(14);
-        expect(updatedState.dailyCounts30Days.length).toBe(30);
-        expect(updatedState.sampledEarthquakesLast14Days).toBeInstanceOf(Array);
-        expect(updatedState.sampledEarthquakesLast30Days).toBeInstanceOf(Array);
-        expect(updatedState.magnitudeDistribution14Days.length).toBe(MAGNITUDE_RANGES.length);
-        expect(updatedState.magnitudeDistribution30Days.length).toBe(MAGNITUDE_RANGES.length);
-    });
-
-    it('should handle empty features for monthly processing', () => {
-      const emptyAction = { type: actionTypes.MONTHLY_DATA_PROCESSED, payload: { features: [], fetchTime: mockFetchTime } };
-      const emptyState = earthquakeReducer(initialState, emptyAction);
-      expect(emptyState.isLoadingMonthly).toBe(false);
-      expect(emptyState.allEarthquakes.length).toBe(0);
-      expect(emptyState.earthquakesLast14Days.length).toBe(0);
-      expect(emptyState.earthquakesLast30Days.length).toBe(0);
-      expect(emptyState.dailyCounts14Days.every(d => d.count === 0)).toBe(true);
-      expect(emptyState.dailyCounts30Days.every(d => d.count === 0)).toBe(true);
-      expect(emptyState.sampledEarthquakesLast14Days.length).toBe(0);
-      expect(emptyState.sampledEarthquakesLast30Days.length).toBe(0);
-      expect(emptyState.magnitudeDistribution14Days.every(d => d.count === 0)).toBe(true);
-      expect(emptyState.magnitudeDistribution30Days.every(d => d.count === 0)).toBe(true);
-      expect(emptyState.lastMajorQuake).toBeNull();
-    });
-
-    it('should clear monthlyError', () => {
-        const stateWithError={...initialState,monthlyError:"err"};
-        const updatedStateAfterErrorClear = earthquakeReducer(stateWithError,action);
-        expect(updatedStateAfterErrorClear.monthlyError).toBeNull();
-    });
-  });
-});
+// MOVED to src/tests/contexts/EarthquakeDataContext.reducer.test.jsx
 
 // --- Tests for Helper Functions (Restored with placeholder .todo for brevity in this example) ---
 describe('Helper: filterByTime', () => { it.todo('tests need to be fully restored for filterByTime'); });
@@ -286,411 +48,10 @@ describe('Helper: calculateMagnitudeDistribution', () => { it.todo('tests need t
 
 
 // --- Tests for loadMonthlyData and Memoized Selectors ---
-const AllTheProviders = ({ children }) => (<EarthquakeDataProvider>{children}</EarthquakeDataProvider>);
+// MOVED to src/tests/contexts/EarthquakeDataContext.provider.test.jsx and src/tests/contexts/EarthquakeDataContext.selectors.test.jsx
 
 // --- Tests for EarthquakeDataProvider async logic and initial load ---
-describe('EarthquakeDataProvider initial load and refresh', () => {
-  let setIntervalSpy;
-  let clearIntervalSpy;
-  let intervalCallbacks = {};
-  let intervalIdCounter = 0;
+// MOVED to src/tests/contexts/EarthquakeDataContext.provider.test.jsx
 
-  beforeEach(() => {
-    vi.useFakeTimers(); // Still use fake timers for other time-based logic if needed
-    fetchUsgsData.mockReset();
-
-    intervalCallbacks = {}; // Reset on each test
-    intervalIdCounter = 0;
-
-    setIntervalSpy = vi.spyOn(global, 'setInterval');
-    clearIntervalSpy = vi.spyOn(global, 'clearInterval');
-
-    setIntervalSpy.mockImplementation((callback, timeout) => {
-      const id = ++intervalIdCounter;
-      // console.log(`Mock setInterval called: id=${id}, timeout=${timeout}`);
-      intervalCallbacks[id] = { callback, timeout, type: timeout === REFRESH_INTERVAL_MS ? 'refresh' : 'loadingMessage' };
-      return id;
-    });
-
-    clearIntervalSpy.mockImplementation((id) => {
-      // console.log(`Mock clearInterval called: id=${id}`);
-      delete intervalCallbacks[id];
-    });
-  });
-
-  afterEach(() => {
-    // Restore original timers and clear any spies
-    // vi.restoreAllMocks() // This would restore all mocks, might be too broad if other spies are used intentionally
-    setIntervalSpy.mockRestore();
-    clearIntervalSpy.mockRestore();
-
-    vi.runOnlyPendingTimers(); // Clear any remaining standard fake timers
-    vi.useRealTimers();
-    vi.clearAllTimers(); // Ensure Vitest's own fake timers are cleared
-    intervalCallbacks = {}; // Clean up
-  });
-
-  // Helper to run specific intervals by type
-  const runIntervals = async (type, runMax = Infinity) => {
-    let runCount = 0;
-    // console.log(`Running intervals of type: ${type}. Found:`, Object.values(intervalCallbacks).filter(ic => ic.type === type).length);
-    for (const id in intervalCallbacks) {
-        if (intervalCallbacks[id].type === type && runCount < runMax) {
-            // console.log(`Manually calling interval id: ${id} of type ${type}`);
-            await act(async () => {
-                intervalCallbacks[id].callback();
-            });
-            runCount++;
-        }
-    }
-  };
-
-  const runAllIntervalsMultipleTimes = async (count = 3) => {
-    for (let i = 0; i < count; i++) {
-        for (const id in intervalCallbacks) {
-             await act(async () => {
-                intervalCallbacks[id].callback();
-            });
-        }
-        await act(async () => { await Promise.resolve(); }); // Flush promises between runs
-    }
-  };
-
-
-  it('should perform initial data load (daily & weekly) on mount and set loading states', async () => {
-    // Renamed these consts to avoid conflict with broader scope variables
-    const specificTest_mockDailyData = { features: [{id: 'd1', properties: {time: Date.now(), mag: 1}}], metadata: { generated: Date.now() }};
-    const specificTest_mockWeeklyData = { features: [{id: 'w1', properties: {time: Date.now(), mag: 2}}], metadata: { generated: Date.now() }};
-
-    let dailyFetchResolved = false;
-    let weeklyFetchResolved = false;
-
-    fetchUsgsData.mockImplementation(async (url) => {
-        await Promise.resolve(); // Simulate async nature of fetch
-        if (url === USGS_API_URL_DAY) {
-            dailyFetchResolved = true;
-            return specificTest_mockDailyData; // Use renamed variable
-        }
-        if (url === USGS_API_URL_WEEK) {
-            weeklyFetchResolved = true;
-            return specificTest_mockWeeklyData; // Use renamed variable
-        }
-        return { features: [], metadata: {} };
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-
-    // Initial state assertions
-    expect(result.current.isLoadingDaily).toBe(true);
-    expect(result.current.isLoadingWeekly).toBe(true);
-    expect(result.current.isInitialAppLoad).toBe(true);
-
-    // Advance timers enough for loading messages to cycle a bit
-    // and for initial effects in orchestrateInitialDataLoad to fire
-    // Initial render (the renderHook above this block is the correct one for this test)
-
-    // Initial state assertions
-    expect(result.current.isLoadingDaily).toBe(true);
-    expect(result.current.isLoadingWeekly).toBe(true);
-    expect(result.current.isInitialAppLoad).toBe(true);
-
-    // Manually trigger loading message intervals a few times
-    await runIntervals('loadingMessage');
-    await runIntervals('loadingMessage');
-
-    // Allow promises from fetchUsgsData (called by orchestrateInitialDataLoad on mount) to resolve
-    await act(async () => {
-      await Promise.resolve(); // Flush microtasks for fetch promises
-      await Promise.resolve(); // Again to be safe for chained promises
-    });
-
-    // After fetches, state updates should have happened
-    expect(dailyFetchResolved).toBe(true); // Check if mock fetch was called and resolved
-    expect(weeklyFetchResolved).toBe(true);
-    expect(fetchUsgsData).toHaveBeenCalledWith(USGS_API_URL_DAY);
-    expect(fetchUsgsData).toHaveBeenCalledWith(USGS_API_URL_WEEK);
-    expect(result.current.isLoadingDaily).toBe(false);
-    expect(result.current.isLoadingWeekly).toBe(false);
-    expect(result.current.isInitialAppLoad).toBe(false);
-    expect(result.current.earthquakesLastHour.length).toBeGreaterThanOrEqual(0);
-    expect(result.current.earthquakesLast7Days.length).toBeGreaterThanOrEqual(0);
-  });
-
-  it('should handle error if daily fetch fails during initial load', async () => {
-    const mockWeeklyData = { features: [{id: 'w1', properties: {time: Date.now(), mag: 1}}], metadata: {generated: Date.now()} }; // Added properties
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_DAY) return Promise.resolve({ error: { message: "Daily fetch failed" } });
-      if (url === USGS_API_URL_WEEK) return Promise.resolve(mockWeeklyData);
-      return Promise.resolve({ features: [] });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-
-    await act(async () => {
-        await Promise.resolve(); // Allow fetches to "complete"
-        await Promise.resolve();
-    });
-    await runAllIntervalsMultipleTimes(2); // Run loading/refresh intervals
-
-    expect(result.current.isLoadingDaily).toBe(false);
-    expect(result.current.isLoadingWeekly).toBe(false);
-    expect(result.current.error).toContain("Daily data error: Daily fetch failed");
-    expect(result.current.isInitialAppLoad).toBe(false);
-  });
-
-  it('should handle error if weekly fetch fails during initial load', async () => {
-    const mockDailyData = { features: [{id: 'd1', properties: {time: Date.now(), mag: 1}}], metadata: {generated: Date.now()} }; // Added properties
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_DAY) return Promise.resolve(mockDailyData);
-      if (url === USGS_API_URL_WEEK) return Promise.resolve({ error: { message: "Weekly fetch failed" } });
-      return Promise.resolve({ features: [] });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-
-    await act(async () => {
-        await Promise.resolve();
-        await Promise.resolve();
-    });
-    await runAllIntervalsMultipleTimes(2);
-
-    expect(result.current.isLoadingDaily).toBe(false);
-    expect(result.current.isLoadingWeekly).toBe(false);
-    expect(result.current.error).toContain("Weekly data error: Weekly fetch failed");
-    expect(result.current.isInitialAppLoad).toBe(false);
-  });
-
-  it('should handle errors if both daily and weekly fetches fail during initial load', async () => {
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_DAY) return Promise.resolve({ error: { message: "Daily failed" } });
-      if (url === USGS_API_URL_WEEK) return Promise.resolve({ error: { message: "Weekly failed" } });
-      return Promise.resolve({ features: [] });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-
-    await act(async () => {
-        await Promise.resolve();
-        await Promise.resolve();
-    });
-    await runAllIntervalsMultipleTimes(2);
-
-    expect(result.current.error).toBe("Failed to fetch critical daily and weekly data.");
-    expect(result.current.isInitialAppLoad).toBe(false);
-  });
-
-  it.skip('should cycle loading messages during initial load and stop after', async () => {
-    // Ensure distinct successful responses for daily and weekly initial loads
-    fetchUsgsData
-        .mockResolvedValueOnce({ features: [{id:'q_daily_cycle_test', properties:{time: Date.now(), mag:1}}], metadata: { generated: Date.now() } }) // For daily
-        .mockResolvedValueOnce({ features: [{id:'q_weekly_cycle_test', properties:{time: Date.now(), mag:2}}], metadata: { generated: Date.now() } }); // For weekly
-
-    const initialMessages = contextInitialState.currentLoadingMessages;
-    expect(initialMessages.length).toBeGreaterThan(1);
-
-    let result;
-
-    // Initial Render + synchronous effects from orchestrateInitialDataLoad
-    await act(async () => {
-        const { result: hookResult } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-        result = hookResult;
-        // Allow microtasks to flush, ensuring initial sync dispatches in useEffect -> orchestrateInitialDataLoad complete
-        // Vitest's act typically handles promise flushing, but an explicit await can sometimes help ensure order.
-        await new Promise(setImmediate); // More robust than setTimeout(0) for promise flushing in some test envs
-    });
-
-    // Check message after the two synchronous UPDATE_LOADING_MESSAGE_INDEX dispatches from orchestrateInitialDataLoad
-    // Initial state: index 0. Dispatch 1: index 1. Dispatch 2: index 2.
-    const expectedMessageAfterSyncDispatches = initialMessages.length >= 3 ? initialMessages[2] :
-                                              (initialMessages.length === 2 ? initialMessages[0] : initialMessages[0]);
-    expect(result.current.currentLoadingMessage).toBe(expectedMessageAfterSyncDispatches);
-    expect(result.current.isInitialAppLoad).toBe(true);
-
-    // Use waitFor to ensure isInitialAppLoad becomes false, indicating loading completion.
-    // The timeout for waitFor should be less than the test's own timeout.
-    // This also allows React to process state updates and effects naturally with fake timers.
-    await act(async () => {
-        // It's important that `fetchUsgsData` mocks resolve.
-        // Advancing timers helps if there are any `setTimeout` or `setInterval` involved in the loading process itself,
-        // beyond the message cycling. The `orchestrateInitialDataLoad` uses async/await and promises.
-        // We need to ensure these promises resolve and their effects on state are processed.
-        // Running all timers might be too aggressive if it clears the message interval prematurely for the test's logic.
-        // Let's advance by a small amount to ensure any initial async setup starts.
-        vi.advanceTimersByTime(1);
-        // The key is waiting for the state to change, not just timers.
-    });
-
-    await waitFor(() => {
-      expect(result.current.isInitialAppLoad).toBe(false);
-    }, { timeout: 4800 }); // Slightly less than the default 5s test timeout. Increased from 4500
-
-    // Verify that fetch was called for both daily and weekly data
-    expect(fetchUsgsData).toHaveBeenCalledWith(USGS_API_URL_DAY);
-    expect(fetchUsgsData).toHaveBeenCalledWith(USGS_API_URL_WEEK);
-    expect(fetchUsgsData).toHaveBeenCalledTimes(2);
-
-    const messageWhenLoadFinished = result.current.currentLoadingMessage;
-
-    // Now that isInitialAppLoad is false, the interval should have been cleared.
-    // Advance time again to check if the message *still* changes. It shouldn't.
-    await act(async () => {
-      vi.advanceTimersByTime(LOADING_MESSAGE_INTERVAL_MS * 3);
-    });
-
-    expect(result.current.currentLoadingMessage).toBe(messageWhenLoadFinished, "Loading message should not change after initial load is complete and interval is cleared.");
-  });
-
-  // Test for refresh logic needs to be adapted for manual interval trigger
-  it('should refresh data when refresh interval callback is manually triggered', async () => {
-    const initialDailyTime = Date.now();
-    const initialWeeklyTime = initialDailyTime - 1000; // Ensure distinct times
-
-    fetchUsgsData.mockResolvedValueOnce({ features: [{id:'q_initial_daily', properties: {time: initialDailyTime, mag: 1}}], metadata: {generated: initialDailyTime} })
-                   .mockResolvedValueOnce({ features: [{id:'q_initial_weekly', properties: {time: initialWeeklyTime, mag: 1}}], metadata: {generated: initialWeeklyTime} });
-
-    let result;
-    await act(async () => {
-        const { result: hookResult } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-        result = hookResult;
-        await Promise.resolve();
-        await Promise.resolve();
-        await runIntervals('loadingMessage', 2); // Settle initial loading messages
-    });
-
-    expect(fetchUsgsData).toHaveBeenCalledTimes(2);
-
-    // Setup for refresh call
-    fetchUsgsData.mockClear();
-    const refreshFetchTime = Date.now() + 5000; // Simulate time passing for the refresh
-    const refreshedDailyQuakeTime = refreshFetchTime - 1000; // 1 second before this "new Date.now()"
-
-    fetchUsgsData.mockResolvedValueOnce({ features: [{id:'q_refresh_daily', properties: {time: refreshedDailyQuakeTime, mag: 2}}], metadata: {generated: refreshFetchTime} })
-                   .mockResolvedValueOnce({ features: [{id:'q_refresh_weekly', properties: {time: refreshedDailyQuakeTime - 2000, mag: 2}}], metadata: {generated: refreshFetchTime} });
-
-    // Manually trigger THE refresh interval's callback
-    // The orchestrateInitialDataLoad called by refresh will use its own Date.now() for filtering.
-    // We need to ensure our mocked feature times are relative to that.
-    // For the test, we'll mock Date.now() just for the duration of the refresh callback.
-    vi.spyOn(global.Date, 'now').mockReturnValueOnce(refreshFetchTime);
-
-    await act(async () => {
-        await runIntervals('refresh', 1);
-        await Promise.resolve();
-        await Promise.resolve();
-    });
-
-    global.Date.now.mockRestore(); // IMPORTANT: Restore Date.now
-
-    expect(fetchUsgsData).toHaveBeenCalledTimes(2);
-    expect(result.current.earthquakesLastHour.some(q => q.id === 'q_refresh_daily')).toBe(true);
-  });
-
-});
-
-
-describe('EarthquakeDataContext: loadMonthlyData', () => {
-  beforeEach(() => {
-    fetchUsgsData.mockReset();
-  });
-
-  it('should fetch monthly data and dispatch MONTHLY_DATA_PROCESSED on success', async () => {
-    const minimalMockFeatures = [{ id: 'mocker1', properties: { time: Date.now(), mag: 3.0 } }];
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_MONTH) return Promise.resolve({ features: minimalMockFeatures });
-      return Promise.resolve({ features: [], metadata: {} });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-    await act(async () => new Promise(resolve => setTimeout(resolve, 0)));
-
-    expect(result.current.monthlyError).toBeNull();
-    await act(async () => { result.current.loadMonthlyData(); });
-
-    expect(fetchUsgsData).toHaveBeenCalledWith(USGS_API_URL_MONTH);
-    expect(result.current.isLoadingMonthly).toBe(false);
-    expect(result.current.hasAttemptedMonthlyLoad).toBe(true);
-    expect(result.current.monthlyError).toBeNull();
-    expect(result.current.allEarthquakes.length).toBe(minimalMockFeatures.length);
-    if (minimalMockFeatures.length > 0) expect(result.current.allEarthquakes[0].id).toBe(minimalMockFeatures[0].id);
-  });
-
-  it('should set monthlyError if API returns an error object', async () => {
-    const errorMessage = "API Error for monthly data";
-    const apiErrorResponse = { error: { message: errorMessage } };
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_MONTH) return Promise.resolve(apiErrorResponse);
-      return Promise.resolve({ features: [], metadata: {} });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-    await act(async () => new Promise(resolve => setTimeout(resolve, 0)));
-
-    await act(async () => { result.current.loadMonthlyData(); });
-    expect(result.current.monthlyError).toBe(errorMessage);
-  });
-
-  it('should set monthlyError if API call throws an error', async () => {
-    const thrownErrorMessage = "Network failure";
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_MONTH) return Promise.reject(new Error(thrownErrorMessage));
-      return Promise.resolve({ features: [], metadata: {} });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-    await act(async () => new Promise(resolve => setTimeout(resolve, 0)));
-
-    await act(async () => { result.current.loadMonthlyData(); });
-    expect(result.current.monthlyError).toContain(`Monthly Data Processing Error: ${thrownErrorMessage}`);
-  });
-
-  it('should set monthlyError if API returns no features or empty features array', async () => {
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_MONTH) return Promise.resolve({ features: [] });
-      return Promise.resolve({ features: [], metadata: {} });
-    });
-
-    const { result } = renderHook(() => useEarthquakeDataState(), { wrapper: AllTheProviders });
-    await act(async () => new Promise(resolve => setTimeout(resolve, 0)));
-
-    await act(async () => { result.current.loadMonthlyData(); });
-    expect(result.current.monthlyError).toBe("Monthly data is unavailable or incomplete.");
-
-    fetchUsgsData.mockImplementation(async (url) => {
-      if (url === USGS_API_URL_MONTH) return Promise.resolve({});
-      return Promise.resolve({ features: [], metadata: {} });
-    });
-
-    await act(async () => { result.current.loadMonthlyData(); });
-    expect(result.current.monthlyError).toBe("Monthly data is unavailable or incomplete.");
-  });
-});
-
-describe('EarthquakeDataContext: Memoized Selectors', () => {
-  const createMockQuake = (id, mag) => ({ id, properties: { mag, time: Date.now() } });
-  const mockQuakes7Days = [ createMockQuake('q7_1', 2.0), createMockQuake('q7_2', 2.5), createMockQuake('q7_3', 3.0), createMockQuake('q7_4', 4.5), createMockQuake('q7_5', 5.0), createMockQuake('q7_null', null), ];
-  const mockQuakes30Days = [ createMockQuake('q30_1', 1.0), createMockQuake('q30_2', 2.49), createMockQuake('q30_3', 2.5), createMockQuake('q30_4', 4.0), createMockQuake('q30_5', 4.49), createMockQuake('q30_6', 4.5), createMockQuake('q30_7', 6.0), createMockQuake('q30_undefined', undefined), ];
-
-  it('should correctly compute memoized selectors based on context state', () => {
-    const TestProvider = ({ children, mockState }) => {
-      const currentMockState = mockState || {};
-      const feelableQuakes7Days_ctx = React.useMemo(() => { const data = currentMockState.earthquakesLast7Days || []; return data.filter( q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= FEELABLE_QUAKE_THRESHOLD ); }, [currentMockState.earthquakesLast7Days]);
-      const significantQuakes7Days_ctx = React.useMemo(() => { const data = currentMockState.earthquakesLast7Days || []; return data.filter( q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= MAJOR_QUAKE_THRESHOLD ); }, [currentMockState.earthquakesLast7Days]);
-      const feelableQuakes30Days_ctx = React.useMemo(() => { const data = currentMockState.allEarthquakes || []; return data.filter( q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= FEELABLE_QUAKE_THRESHOLD ); }, [currentMockState.allEarthquakes]);
-      const significantQuakes30Days_ctx = React.useMemo(() => { const data = currentMockState.allEarthquakes || []; return data.filter( q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= MAJOR_QUAKE_THRESHOLD ); }, [currentMockState.allEarthquakes]);
-      const contextValueForTestProvider = { ...currentMockState, feelableQuakes7Days_ctx, significantQuakes7Days_ctx, feelableQuakes30Days_ctx, significantQuakes30Days_ctx, };
-      return <EarthquakeDataContext.Provider value={contextValueForTestProvider}>{children}</EarthquakeDataContext.Provider>;
-    };
-    const mockProviderState = { ...contextInitialState, earthquakesLast7Days: mockQuakes7Days, allEarthquakes: mockQuakes30Days, };
-    const wrapper = ({ children }) => <TestProvider mockState={mockProviderState}>{children}</TestProvider>;
-    const { result: contextValueHookResult } = renderHook(() => useEarthquakeDataState(), { wrapper });
-
-    const expectedFeelable7Days = mockQuakes7Days.filter(q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= FEELABLE_QUAKE_THRESHOLD);
-    expect(contextValueHookResult.current.feelableQuakes7Days_ctx.map(q => q.id).sort()).toEqual(expectedFeelable7Days.map(q => q.id).sort());
-    const expectedSignificant7Days = mockQuakes7Days.filter(q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= MAJOR_QUAKE_THRESHOLD);
-    expect(contextValueHookResult.current.significantQuakes7Days_ctx.map(q => q.id).sort()).toEqual(expectedSignificant7Days.map(q => q.id).sort());
-    const expectedFeelable30Days = mockQuakes30Days.filter(q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= FEELABLE_QUAKE_THRESHOLD);
-    expect(contextValueHookResult.current.feelableQuakes30Days_ctx.map(q => q.id).sort()).toEqual(expectedFeelable30Days.map(q => q.id).sort());
-    const expectedSignificant30Days = mockQuakes30Days.filter(q => q.properties && q.properties.mag !== null && typeof q.properties.mag === 'number' && q.properties.mag >= MAJOR_QUAKE_THRESHOLD);
-    expect(contextValueHookResult.current.significantQuakes30Days_ctx.map(q => q.id).sort()).toEqual(expectedSignificant30Days.map(q => q.id).sort());
-  });
-});
+// --- Tests for Memoized Selectors --- (This was a duplicate describe block, main one moved)
+// MOVED to src/tests/contexts/EarthquakeDataContext.selectors.test.jsx
